+++ bangazon/bangazonapi/views/__init__.py
@@ -1,0 +1,1 @@
+__all__ = ["customer_view","payment_view", 'product_view', 'order_view', 'producttype_view']
