+++ bangazon/bangazonapi/views/product_view.py
@@ -1,0 +1,19 @@
+from rest_framework import viewsets
+from bangazonapi.models import *
+from bangazonapi.serializers import *
+from rest_framework.permissions import *
+from django_filters.rest_framework import DjangoFilterBackend
+from django_filters import rest_framework as filters
+from rest_framework.authentication import SessionAuthentication, BasicAuthentication, TokenAuthentication
+
+class ProductViewSet(viewsets.ModelViewSet):
+    ''' The ProductsViewSet class is a view that lists out all products and details about a product.
+
+    Argument List:
+        -generics.ModelViewSet: This method returns both a collect of data and specifi data from the API.
+
+    Author: Nathan Baker, Python Ponies
+    '''
+    # permission_classes = (IsAdminUser,)
+    queryset = products_model.Product.objects.all()
+    serializer_class = product_serializer.ProductSerializer
